// apps/pilote-ppg/src/client/components/_commons/Sélecteur/Sélecteur.interface.ts
import {
  FieldError,
  FieldErrorsImpl,
  Merge,
  UseFormRegisterReturn,
} from "react-hook-form";

export type SélecteurOption<T> = {
  libellé: string;
  valeur: T;
  désactivée?: boolean;
  cachée?: boolean;
};

export default interface SélecteurProps<T> {
  htmlName: string;
  options: SélecteurOption<T>[];
  onChange?: (valeur: T) => void;
  valeurSélectionnée?: T;
  texteAide?: string;
  erreur?: FieldError | Merge<FieldError, FieldErrorsImpl>;
  hasErrors?: boolean;
  errorMessage?: string;
  libellé?: string;
  texteFantôme?: string;
  register?: UseFormRegisterReturn;
  estDesactive?: boolean;
}